Update to adopt uxarray v2023.06
Uxarray v2023.06 brought in significant changes in the data structure,
and rather large changes are made to adopt the new version of uxarray.
For now, all non-default names are swapped to use a default `Mesh2`
instead of SCHISM mesh names. This change is not ideal, but the easiest
(and probably the only way) to use the uxarray for now.

SCH-619  GH-2